Refactor RequirementsList
